+++ tests/unit/hive_metastore/test_migrate.py
@@ -1,8 +1,7 @@
 import logging
 from unittest.mock import MagicMock
 
-from databricks.labs.ucx.hive_metastore.migrate_tables import TablesMigrate
-from databricks.labs.ucx.hive_metastore.tables import TablesCrawler
+from databricks.labs.ucx.hive_metastore.tables import TablesCrawler, TablesMigrate
 
 from ..framework.mocks import MockBackend
 
@@ -28,7 +27,7 @@ def test_migrate_managed_tables_should_produce_proper_queries():
     backend = MockBackend(fails_on_first=errors, rows=rows)
     tc = TablesCrawler(backend, "inventory_database")
     client = MagicMock()
-    tm = TablesMigrate(tc, client, backend)
+    tm = TablesMigrate(tc, client, backend, "")
     tm.migrate_tables()
 
     assert (list(backend.queries)) == [
@@ -49,7 +48,7 @@ def test_migrate_managed_tables_should_do_nothing_if_upgrade_tag_is_present():
     backend = MockBackend(fails_on_first=errors, rows=rows)
     tc = TablesCrawler(backend, "inventory_database")
     client = MagicMock()
-    tm = TablesMigrate(tc, client, backend)
+    tm = TablesMigrate(tc, client, backend, "")
     tm.migrate_tables()
 
     assert (list(backend.queries)) == ["SELECT * FROM hive_metastore.inventory_database.tables"]
@@ -66,7 +65,7 @@ def test_migrate_tables_should_migrate_tables_to_default_catalog_if_not_found_in
     tc = TablesCrawler(backend, "inventory_database")
     client = MagicMock()
     database_to_catalog_mapping = {"db1": "catalog_1", "db2": "catalog_2"}
-    tm = TablesMigrate(tc, client, backend, database_to_catalog_mapping=database_to_catalog_mapping)
+    tm = TablesMigrate(tc, client, backend, "", database_to_catalog_mapping=database_to_catalog_mapping)
     tm.migrate_tables()
 
     assert (list(backend.queries)) == [
@@ -87,7 +86,7 @@ def test_migrate_tables_should_migrate_tables_to_default_catalog_if_specified():
     backend = MockBackend(fails_on_first=errors, rows=rows)
     tc = TablesCrawler(backend, "inventory_database")
     client = MagicMock()
-    tm = TablesMigrate(tc, client, backend, default_catalog="test_catalog")
+    tm = TablesMigrate(tc, client, backend, "", default_catalog="test_catalog")
     tm.migrate_tables()
 
     assert (list(backend.queries)) == [
